+++ lib/parse/query/constraint.rb
@@ -107,7 +107,7 @@ module Parse
         d = { __type: Parse::Model::TYPE_DATE, iso: d.iso8601(3) } if d.respond_to?(:iso8601)
         d = d.pointer if d.respond_to?(:pointer) #simplified query object
         d = d.to_s if d.is_a?(Regexp)
-        #d = d.pointer if d.is_a?(Parse::Object) #simplified query object
+        # d = d.pointer if d.is_a?(Parse::Object) #simplified query object
         #  d = d.compile
         if d.is_a?(Parse::Query)
           compiled = d.compile(encode: false, includeClassName: true)